fix: refactor into item types
