+++ package.json
@@ -35,7 +35,7 @@
     "hooks": "husky"
   },
   "dependencies": {
-    "@faker-js/faker": "8.4.0",
+    "@faker-js/faker": "8.4.1",
     "js-cookie": "3.0.5",
     "validator": "13.11.0"
   },
@@ -44,26 +44,26 @@
     "uuid": "^9"
   },
   "devDependencies": {
-    "@commitlint/cli": "18.6.0",
-    "@commitlint/config-conventional": "18.6.0",
+    "@commitlint/cli": "18.6.1",
+    "@commitlint/config-conventional": "18.6.2",
     "@trivago/prettier-plugin-sort-imports": "4.3.0",
     "@types/eslint": "8.56.2",
     "@types/js-cookie": "3.0.6",
     "@types/uuid": "9.0.8",
-    "@types/validator": "13.11.8",
-    "@typescript-eslint/eslint-plugin": "6.20.0",
-    "@typescript-eslint/parser": "6.20.0",
+    "@types/validator": "13.11.9",
+    "@typescript-eslint/eslint-plugin": "7.0.2",
+    "@typescript-eslint/parser": "7.0.2",
     "date-fns": "3.3.1",
     "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.0",
-    "husky": "9.0.7",
+    "husky": "9.0.11",
     "jsdom": "24.0.0",
-    "prettier": "3.2.4",
+    "prettier": "3.2.5",
     "tsc-alias": "1.8.8",
     "typescript": "5.3.3",
     "uuid": "9.0.1",
-    "vite-plugin-dts": "3.7.2",
-    "vitest": "1.2.2"
+    "vite-plugin-dts": "3.7.3",
+    "vitest": "1.3.1"
   },
   "packageManager": "yarn@4.1.0"
 }
