--- src/enums/maxWidth.ts
@@ -1,6 +1,0 @@
-export enum MaxWidth {
-  Small = 'sm',
-  Medium = 'md',
-  Large = 'lg',
-  ExtraLarge = 'xl',
-}
