+++ src/item/factory.ts
@@ -1,22 +1,10 @@
 import { MemberFactory } from '../member/factory.js';
-import {
-  AppItemType,
-  DiscriminatedItem,
-  DocumentItemType,
-  EmbeddedLinkItemType,
-  EtherpadItemType,
-  FolderItemType,
-  H5PItemType,
-  LocalFileItemType,
-  S3FileItemType,
-  ShortcutItemType,
-} from './item.js';
+import { DiscriminatedItem } from './item.js';
 import { buildPathFromIds } from './itemUtils.js';
 import { CCLicenseAdaptions } from '@/enums/ccLicenses.js';
-import { ItemType } from '@/item/itemType.js';
 import { faker } from '@faker-js/faker';
 
-type ItemFactoryOutputType<IT extends DiscriminatedItem> = Pick<
+export type ItemFactoryOutputType<IT extends DiscriminatedItem> = Pick<
   IT,
   | 'id'
   | 'name'
@@ -31,11 +19,11 @@ type ItemFactoryOutputType<IT extends DiscriminatedItem> = Pick<
   | 'updatedAt'
 >;
 
-type ItemFactoryInputType<IT extends DiscriminatedItem> = Partial<IT> & {
+export type ItemFactoryInputType<IT extends DiscriminatedItem> = Partial<IT> & {
   parentItem?: Pick<IT, 'path'>;
 };
 
-const PartialItemFactory = <IT extends DiscriminatedItem>(
+export const PartialItemFactory = <IT extends DiscriminatedItem>(
   item: ItemFactoryInputType<IT> = {},
 ): Omit<ItemFactoryOutputType<IT>, 'type' | 'extra'> => {
   const id = item.id ?? faker.string.uuid();
@@ -73,177 +61,5 @@ const PartialItemFactory = <IT extends DiscriminatedItem>(
     // allow null creator
     creator: item.creator === undefined ? MemberFactory() : item.creator,
     lang: item.lang ?? faker.helpers.arrayElement(['fr', 'en']),
-  };
-};
-
-export const FolderItemFactory = (
-  item: ItemFactoryInputType<FolderItemType> = {},
-): ItemFactoryOutputType<FolderItemType> => {
-  const newItem = PartialItemFactory({ ...item });
-  return {
-    ...newItem,
-    type: ItemType.FOLDER,
-    extra: item.extra ?? {
-      [ItemType.FOLDER]: {
-        childrenOrder: [],
-      },
-    },
-  };
-};
-
-export const AppItemFactory = (
-  item: ItemFactoryInputType<AppItemType> = {},
-): ItemFactoryOutputType<AppItemType> => {
-  const newItem = PartialItemFactory({ ...item, type: ItemType.APP });
-  return {
-    ...newItem,
-    type: ItemType.APP,
-    extra: item.extra ?? {
-      [ItemType.APP]: {
-        url: faker.internet.url(),
-        settings: {},
-      },
-    },
-  };
-};
-
-export const H5PItemFactory = (
-  item: ItemFactoryInputType<H5PItemType> = {},
-): ItemFactoryOutputType<H5PItemType> => {
-  const newItem = PartialItemFactory<H5PItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.H5P,
-    extra: item.extra ?? {
-      [ItemType.H5P]: {
-        contentId: faker.string.uuid(),
-        h5pFilePath: faker.system.filePath(),
-        contentFilePath: faker.system.filePath(),
-      },
-    },
-  };
-};
-
-export const DocumentItemFactory = (
-  item: ItemFactoryInputType<DocumentItemType> = {},
-): ItemFactoryOutputType<DocumentItemType> => {
-  const newItem = PartialItemFactory<DocumentItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.DOCUMENT,
-    extra: item.extra ?? {
-      [ItemType.DOCUMENT]: {
-        content: `<div>${faker.lorem.text()}</div>`,
-      },
-    },
-  };
-};
-
-export const EmbeddedLinkItemFactory = (
-  item: ItemFactoryInputType<EmbeddedLinkItemType> = {},
-): ItemFactoryOutputType<EmbeddedLinkItemType> => {
-  const newItem = PartialItemFactory<EmbeddedLinkItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.LINK,
-    extra: item.extra ?? {
-      [ItemType.LINK]: {
-        html: faker.helpers.arrayElement([
-          `<div>${faker.lorem.text()}</div>`,
-          undefined,
-        ]),
-        icons: faker.helpers.arrayElement([[faker.internet.url()], undefined]),
-        thumbnails: faker.helpers.arrayElement([
-          [faker.internet.url()],
-          undefined,
-        ]),
-        url: faker.internet.url(),
-      },
-    },
-  };
-};
-
-export const LocalFileItemFactory = (
-  item: ItemFactoryInputType<LocalFileItemType> = {},
-): ItemFactoryOutputType<LocalFileItemType> => {
-  const newItem = PartialItemFactory<LocalFileItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.LOCAL_FILE,
-    extra: item.extra ?? {
-      [ItemType.LOCAL_FILE]: {
-        name: faker.system.fileName(),
-        mimetype: faker.system.mimeType(),
-        path: faker.system.filePath(),
-        size: faker.number.int({ max: 100, min: 1 }),
-        altText: faker.helpers.arrayElement([undefined, faker.lorem.word()]),
-        content: faker.lorem.text(),
-      },
-    },
-  };
-};
-
-export const S3FileItemFactory = (
-  item: ItemFactoryInputType<S3FileItemType> = {},
-): ItemFactoryOutputType<S3FileItemType> => {
-  const newItem = PartialItemFactory<S3FileItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.S3_FILE,
-    extra: item.extra ?? {
-      [ItemType.S3_FILE]: {
-        name: faker.system.fileName(),
-        mimetype: faker.system.mimeType(),
-        path: faker.system.filePath(),
-        size: faker.number.int({ max: 100, min: 1 }),
-        altText: faker.helpers.arrayElement([undefined, faker.lorem.word()]),
-        content: faker.lorem.text(),
-      },
-    },
-  };
-};
-
-export const EtherpadItemFactory = (
-  item: ItemFactoryInputType<EtherpadItemType> = {},
-): ItemFactoryOutputType<EtherpadItemType> => {
-  const newItem = PartialItemFactory<EtherpadItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.ETHERPAD,
-    extra: item.extra ?? {
-      [ItemType.ETHERPAD]: {
-        padID: faker.string.uuid(),
-        groupID: faker.string.uuid(),
-      },
-    },
-  };
-};
-
-export const ShortcutItemFactory = (
-  item: ItemFactoryInputType<ShortcutItemType> = {},
-): ItemFactoryOutputType<ShortcutItemType> => {
-  const newItem = PartialItemFactory<ShortcutItemType>({
-    ...item,
-  });
-  return {
-    ...newItem,
-    type: ItemType.SHORTCUT,
-    extra: item.extra ?? {
-      [ItemType.SHORTCUT]: {
-        target: faker.string.uuid(),
-      },
-    },
   };
 };
