+++ src/item/itemSettings.ts
@@ -2,7 +2,6 @@ import {
   CCLicenseAdaptions,
   OldCCLicenseAdaptations,
 } from '@/enums/ccLicenses.js';
-import { MaxWidth } from '@/enums/maxWidth.js';
 
 export interface ItemSettings {
   /** @deprecated use item.lang */
@@ -20,13 +19,4 @@ export interface ItemSettings {
     | CCLicenseAdaptions
     // TODO: these are the old licenses, we might remove them at some point.
     | `${OldCCLicenseAdaptations}`;
-}
-
-export interface EmbeddedLinkItemSettings extends ItemSettings {
-  showLinkIframe?: boolean;
-  showLinkButton?: boolean;
-}
-
-export interface FileItemSettings extends ItemSettings {
-  maxWidth?: MaxWidth;
 }
