+++ tsconfig.json
@@ -20,7 +20,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
 
-    "types": ["vite/client"],
+    "types": ["vite/client", "vitest"],
     "paths": {
       "@/*": ["./src/*"],
     },
